feat: add company folder
