+++ src/pages/api/v1/company/[id]/client/index.ts
@@ -1,63 +1,51 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import version from '@/lib/version';
 import { errorMessageToErrorCode } from '@/lib/utils/errorCode';
-import type { ResponseData } from '../../../../type/iresponsedata';
+import type { ResponseData } from '../../../../../../type/iresponsedata';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<ResponseData>) {
   try {
-    // Info: (20240419 - Jacky) P010001 - GET /payment
     if (!req.headers.userId) {
       throw new Error('Resource not found');
     }
+    // Info: (20240419 - Jacky) C010001 - GET /client
     if (req.method === 'GET') {
-      const paymentList = [
+      const clientList = [
         {
           id: '1',
-          type: 'VISA',
-          no: '1234-1234-1234-1234',
-          expireYear: '29',
-          expireMonth: '01',
-          cvc: '330',
-          name: 'Taiwan Bank',
+          name: 'cafeca',
+          code: '1234',
+          favorite: false,
         },
         {
           id: '2',
-          type: 'VISA',
-          no: '5678-5678-5678-5678',
-          expireYear: '29',
-          expireMonth: '01',
-          cvc: '355',
-          name: 'Taishin International Bank',
+          name: 'isunfa',
+          code: '3333',
+          favorite: false,
         },
       ];
       res.status(200).json({
         powerby: 'ISunFa api ' + version,
         success: true,
         code: '200',
-        message: 'list all payments',
-        payload: paymentList,
+        message: 'list all clients',
+        payload: clientList,
       });
-      // Info: (20240419 - Jacky) P010003 - POST /payment
+      // Info: (20240419 - Jacky) C010003 - POST /client
     } else if (req.method === 'POST') {
-      const { type, no, expireYear, expireMonth, cvc, name } = req.body;
-      if (!type || !no || !expireYear || !expireMonth || !cvc || !name) {
-        throw new Error('Invalid input parameter');
-      }
-      const newPayment = {
+      const { name, code } = req.body;
+      const newClient = {
         id: '3',
-        type,
-        no,
-        expireYear,
-        expireMonth,
-        cvc,
         name,
+        code,
+        favorite: false,
       };
       res.status(200).json({
         powerby: 'ISunFa api ' + version,
         success: true,
         code: '200',
-        message: 'create payment',
-        payload: newPayment,
+        message: 'create client',
+        payload: newClient,
       });
     } else {
       throw new Error('Method Not Allowed');
