+++ src/pages/api/v1/company/[id]/kyc/entity.ts
