+++ src/pages/api/v1/company/[id]/ocr/[resultId]/process_statue.ts
