+++ src/pages/api/v1/company/[id]/ocr/[resultId]/result.ts
