+++ src/pages/api/v1/company/[id]/payment/[id]/index.ts
