+++ src/pages/api/v1/company/[id]/payment/index.ts
@@ -1,56 +1,63 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import version from '@/lib/version';
 import { errorMessageToErrorCode } from '@/lib/utils/errorCode';
-import type { ResponseData } from '../../../../type/iresponsedata';
+import type { ResponseData } from '../../../../../../type/iresponsedata';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<ResponseData>) {
   try {
+    // Info: (20240419 - Jacky) P010001 - GET /payment
     if (!req.headers.userId) {
       throw new Error('Resource not found');
     }
-    // Info: (20240419 - Jacky) S010001 - GET /subscription
     if (req.method === 'GET') {
-      const subscriptionList = [
+      const paymentList = [
         {
           id: '1',
-          entity: 'mermer',
-          plan: 'pro',
-          paymentId: '1',
-          price: 'USD 10',
-          autoRenew: true,
-          expireDate: '2024-01-01',
-          status: 'paid',
+          type: 'VISA',
+          no: '1234-1234-1234-1234',
+          expireYear: '29',
+          expireMonth: '01',
+          cvc: '330',
+          name: 'Taiwan Bank',
+        },
+        {
+          id: '2',
+          type: 'VISA',
+          no: '5678-5678-5678-5678',
+          expireYear: '29',
+          expireMonth: '01',
+          cvc: '355',
+          name: 'Taishin International Bank',
         },
       ];
       res.status(200).json({
         powerby: 'ISunFa api ' + version,
         success: true,
         code: '200',
-        message: 'list all subscriptions',
-        payload: subscriptionList,
+        message: 'list all payments',
+        payload: paymentList,
       });
-      // Info: (20240419 - Jacky) S010002 - POST /subscription
+      // Info: (20240419 - Jacky) P010003 - POST /payment
     } else if (req.method === 'POST') {
-      const { plan, paymentId, autoRenew } = req.body;
-      if (!plan || !paymentId || !autoRenew) {
+      const { type, no, expireYear, expireMonth, cvc, name } = req.body;
+      if (!type || !no || !expireYear || !expireMonth || !cvc || !name) {
         throw new Error('Invalid input parameter');
       }
-      const newSubscription = {
+      const newPayment = {
         id: '3',
-        entity: 'mermer',
-        plan,
-        paymentId,
-        price: 'USD 10',
-        autoRenew,
-        expireDate: '2024-01-01',
-        status: 'paid',
+        type,
+        no,
+        expireYear,
+        expireMonth,
+        cvc,
+        name,
       };
       res.status(200).json({
         powerby: 'ISunFa api ' + version,
         success: true,
         code: '200',
-        message: 'create subscription',
-        payload: newSubscription,
+        message: 'create payment',
+        payload: newPayment,
       });
     } else {
       throw new Error('Method Not Allowed');
