+++ src/pages/api/v1/company/[id]/subscription/index.test.ts
