+++ src/pages/api/v1/company/[id]/vouchers/[resultId]/index.ts
